+++ R/bru.inference.R
@@ -19,367 +19,6 @@ generate <- function(object, ...) {
   UseMethod("generate")
 }
 
-bru_check_object_bru <- function(object,
-                                 new_version = getNamespaceVersion("inlabru")) {
-  object <-
-    bru_info_upgrade(
-      object,
-      new_version = new_version
-    )
-  object
-}
-
-bru_info_upgrade <- function(object,
-                             new_version = getNamespaceVersion("inlabru")) {
-  object_full <- object
-  object <- object[["bru_info"]]
-  msg <- NULL
-  if (!is.list(object)) {
-    msg <- "Not a list"
-  } else if (!inherits(object, "bru_info")) {
-    msg <- "Not a bru_info object"
-  } else if (is.null(object[["inlabru_version"]])) {
-    msg <- "`inlabru_version` is missing"
-  }
-  if (!is.null(msg)) {
-    stop(glue(
-      "bru_info part of the object can't be converted to `bru_info`; {msg}"
-    ))
-  }
-
-  old_ver <- object[["inlabru_version"]]
-  if (utils::compareVersion(new_version, old_ver) > 0) {
-    warning(
-      glue("Old bru_info object version {old_ver} detected.
-            Attempting upgrade to version {new_version}.")
-    )
-
-    message("Detected bru_info version ", old_ver)
-    if (utils::compareVersion("2.1.14.901", old_ver) > 0) {
-      message("Upgrading bru_info to 2.1.14.901")
-      # Ensure INLA_version stored
-      if (is.null(object[["INLA_version"]])) {
-        object[["INLA_version"]] <- "0.0.0"
-      }
-      # Check for component$mapper as a bm_multi
-      for (k in seq_along(object[["model"]][["effects"]])) {
-        cmp <- object[["model"]][["effects"]][[k]]
-        cmp[["mapper"]] <-
-          bm_multi(
-            list(
-              main = cmp$main$mapper,
-              group = cmp$group$mapper,
-              replicate = cmp$replicate$mapper
-            ),
-            simplify = TRUE
-          )
-        object[["model"]][["effects"]][[k]] <- cmp
-      }
-      object[["inlabru_version"]] <- "2.1.14.901"
-    }
-    if (utils::compareVersion("2.5.3.9003", old_ver) > 0) {
-      message("Upgrading bru_info to 2.5.3.9003")
-      # Check that likelihoods store 'weights'
-      for (k in seq_along(object[["lhoods"]])) {
-        lhood <- object[["lhoods"]][[k]]
-        if (is.null(lhood[["weights"]])) {
-          lhood[["weights"]] <- 1
-          object[["lhoods"]][[k]] <- lhood
-        }
-      }
-      object[["inlabru_version"]] <- "2.5.3.9003"
-    }
-
-    if (utils::compareVersion("2.5.3.9005", old_ver) > 0) {
-      message("Upgrading bru_info to 2.5.3.9005")
-      # Make sure component$mapper is a bm_pipe
-      for (k in seq_along(object[["model"]][["effects"]])) {
-        cmp <- object[["model"]][["effects"]][[k]]
-        # Convert offset mappers to const mappers
-        if (inherits(cmp$main$mapper, c("bm_offset", "bru_mapper_offset"))) {
-          cmp$main$mapper <- bm_const()
-        }
-        cmp[["mapper"]] <-
-          bm_pipe(
-            list(
-              mapper = bm_multi(
-                list(
-                  main = cmp$main$mapper,
-                  group = cmp$group$mapper,
-                  replicate = cmp$replicate$mapper
-                ),
-                simplify = TRUE
-              ),
-              scale = bm_scale()
-            )
-          )
-        object[["model"]][["effects"]][[k]] <- cmp
-      }
-      object[["inlabru_version"]] <- "2.5.3.9005"
-    }
-
-    if (utils::compareVersion("2.6.0.9000", old_ver) > 0) {
-      message("Upgrading bru_info to 2.6.0.9000")
-      # Make sure component$mapper is a bm_pipe
-      for (k in seq_along(object[["model"]][["effects"]])) {
-        cmp <- object[["model"]][["effects"]][[k]]
-        cmp[["mapper"]] <-
-          bm_pipe(
-            list(
-              mapper = bm_multi(
-                list(
-                  main = cmp$main$mapper,
-                  group = cmp$group$mapper,
-                  replicate = cmp$replicate$mapper
-                ),
-                simplify = TRUE
-              ),
-              scale = bm_scale()
-            )
-          )
-        object[["model"]][["effects"]][[k]] <- cmp
-      }
-      object[["inlabru_version"]] <- "2.6.0.9000"
-    }
-
-    if (utils::compareVersion("2.7.0.9010", old_ver) > 0) {
-      message("Upgrading bru_info to 2.7.0.9010")
-      # Make sure component$group/replicate$input isn't NULL
-      for (k in seq_along(object[["model"]][["effects"]])) {
-        cmp <- object[["model"]][["effects"]][[k]]
-        if (identical(deparse(cmp[["group"]][["input"]][["input"]]), "NULL")) {
-          cmp[["group"]][["input"]][["input"]] <- expression(1L)
-        }
-        if (identical(
-          deparse(cmp[["replicate"]][["input"]][["input"]]),
-          "NULL"
-        )) {
-          cmp[["replicate"]][["input"]][["input"]] <- expression(1L)
-        }
-        object[["model"]][["effects"]][[k]] <- cmp
-      }
-      object[["inlabru_version"]] <- "2.7.0.9010"
-    }
-
-    if (utils::compareVersion("2.7.0.9016", old_ver) > 0) {
-      message("Upgrading bru_info to 2.7.0.9016")
-      # Convert old style include_component/allow_latent to intermediate
-      # include_component/include_latent format
-
-      # Update include/exclude information to limit it to existing components
-      for (k in seq_along(object[["lhoods"]])) {
-        if (isTRUE(object[["lhoods"]][[k]][["allow_latent"]])) {
-          # Force inclusion of all components
-          object[["lhoods"]][[k]][["include_latent"]] <- NULL
-        } else {
-          if (is.null(object[["lhoods"]][[k]][["include_latent"]])) {
-            object[["lhoods"]][[k]][["include_latent"]] <- character(0)
-          }
-        }
-      }
-      object[["lhoods"]] <-
-        bru_used_upgrade(
-          object[["lhoods"]],
-          labels = names(object[["model"]][["effects"]])
-        )
-      object[["inlabru_version"]] <- "2.7.0.9016"
-    }
-
-    if (utils::compareVersion("2.7.0.9017", old_ver) > 0) {
-      message("Upgrading bru_info to 2.7.0.9017")
-      # Convert old style include_component/allow_latent to new
-      # bru_used format
-
-      # Update include/exclude information to limit it to existing components
-      for (k in seq_along(object[["lhoods"]])) {
-        if (is.null(object[["lhoods"]][[k]][["used_components"]])) {
-          if (isTRUE(object[["lhoods"]][[k]][["allow_latent"]])) {
-            # Force inclusion of all components
-            object[["lhoods"]][[k]][["include_latent"]] <- NULL
-          } else {
-            if (is.null(object[["lhoods"]][[k]][["include_latent"]])) {
-              object[["lhoods"]][[k]][["include_latent"]] <- character(0)
-            }
-          }
-        }
-      }
-      object[["lhoods"]] <-
-        bru_used_upgrade(
-          object[["lhoods"]],
-          labels = names(object[["model"]][["effects"]])
-        )
-      object[["inlabru_version"]] <- "2.7.0.9017"
-    }
-
-    if (utils::compareVersion("2.7.0.9021", old_ver) > 0) {
-      message("Upgrading bru_info to 2.7.0.9021")
-      # Make sure 'used' components format is properly stored
-
-      if (is.null(object[["lhoods"]][["is_additive"]])) {
-        object[["lhoods"]][["is_additive"]] <-
-          object[["lhoods"]][["linear"]]
-      }
-
-      object[["lhoods"]] <-
-        bru_used_upgrade(
-          object[["lhoods"]],
-          labels = names(object[["model"]][["effects"]])
-        )
-      object[["inlabru_version"]] <- "2.7.0.9021"
-    }
-
-    if (utils::compareVersion("2.10.1.9007", old_ver) > 0) {
-      message("Upgrading bru_info to 2.10.1.9007")
-      # Update timings info to difftime format
-
-      warning(
-        glue(
-          "
-           From 2.10.1.9007, elapsed time is in Elapsed, Time is CPU time.
-             Copying old elapsed time to both Elapsed and Time,
-             setting System to zero;
-             Do not over-interpret."
-        ),
-        immediate. = TRUE
-      )
-      conversion <- as.difftime(
-        object_full[["bru_timings"]][["Time"]],
-        units = "secs"
-      )
-      object_full[["bru_timings"]][["Time"]] <- conversion
-      object_full[["bru_timings"]][["System"]] <-
-        as.difftime(rep(0.0, length(conversion)), units = "secs")
-      object_full[["bru_timings"]][["Elapsed"]] <- conversion
-
-      object[["inlabru_version"]] <- "2.10.1.9007"
-    }
-
-    if (utils::compareVersion("2.10.1.9012", old_ver) > 0) {
-      message("Upgrading bru_info to 2.10.1.9012")
-      # Update log format to new format
-
-      object_full[["bru_iinla"]][["log"]] <-
-        bru_log_new(
-          object_full[["bru_iinla"]][["log"]][["log"]],
-          bookmarks = object_full[["bru_iinla"]][["log"]][["bookmarks"]]
-        )
-
-      object[["inlabru_version"]] <- "2.10.1.9012"
-    }
-
-    if (utils::compareVersion("2.12.0.9014", old_ver) > 0) {
-      message("Upgrading bru_info to 2.12.0.9014")
-      # Update is_additive/linear storage
-
-      object[["lhoods"]][["is_additive"]] <-
-        object[["lhoods"]][["linear"]]
-
-      # Update predictor expression
-      object[["lhoods"]] <-
-        bru_used_upgrade(
-          object[["lhoods"]],
-          labels = names(object[["model"]][["effects"]])
-        )
-
-      object[["inlabru_version"]] <- "2.12.0.9014"
-    }
-
-    if (utils::compareVersion("2.12.0.9017", old_ver) > 0) {
-      message("Upgrading bru_info to 2.12.0.9017")
-
-      # Update bru_like class names to bru_obs
-      if (!is.null(object[["lhoods"]])) {
-        object[["lhoods"]] <-
-          lapply(object[["lhoods"]], function(x) {
-            class(x) <- "bru_obs"
-            x
-          })
-        class(object[["lhoods"]]) <- c("bru_obs_list", "list")
-      }
-
-      eff <- object[["model"]][["effects"]]
-      for (k in seq_along(object[["model"]][["effects"]])) {
-        class(eff[[k]][["main"]]) <- "bru_subcomp"
-        class(eff[[k]][["group"]]) <- "bru_subcomp"
-        class(eff[[k]][["replicate"]]) <- "bru_subcomp"
-        class(eff[[k]]) <- "bru_comp"
-      }
-      class(eff) <- c("bru_comp_list", "list")
-      eff <- object[["model"]][["effects"]] <- eff
-
-      object[["inlabru_version"]] <- "2.12.0.9017"
-    }
-
-    if (utils::compareVersion("2.13.0.9012", old_ver) > 0) {
-      message("Upgrading bru_info to 2.13.0.9012")
-
-      # Update bru_input input&layer to quosures
-      quosure_update <- function(x, env) {
-        if (is.null(x)) {
-          return(NULL)
-        }
-        rlang::as_quosure(x, env = env)
-      }
-      eff <- object[["model"]][["effects"]]
-      for (k in seq_along(object[["model"]][["effects"]])) {
-        for (part in c("main", "group", "replicate")) {
-          if (!is.null(eff[[k]][[part]][["input"]])) {
-            for (input_part in c("input", "layer")) {
-              eff[[k]][[part]][["input"]][[input_part]] <-
-                quosure_update(
-                  eff[[k]][[part]][["input"]][[input_part]],
-                  eff[[k]][["env_extra"]]
-                )
-            }
-          }
-        }
-        part <- "weights"
-        if (!is.null(eff[[k]][[part]])) {
-          for (input_part in c("input", "layer")) {
-            eff[[k]][[part]][[input_part]] <-
-              quosure_update(
-                eff[[k]][[part]][[input_part]],
-                eff[[k]][["env_extra"]]
-              )
-          }
-        }
-      }
-      eff <- object[["model"]][["effects"]] <- eff
-
-      object[["inlabru_version"]] <- "2.13.0.9012"
-    }
-
-    if (utils::compareVersion("2.13.0.9015", old_ver) > 0) {
-      message("Upgrading bru_info to 2.13.0.9015")
-
-      # Update lhoods to streamlined storage format
-      if (!is.null(object[["lhoods"]])) {
-        object[["lhoods"]] <-
-          lapply(object[["lhoods"]], function(x) {
-            for (what in c("E", "Ntrials", "weights", "scale")) {
-              BRU_what <- paste0("BRU_", what)
-              if (is.null(x[["response_data"]][[BRU_what]])) {
-                x[["response_data"]][[BRU_what]] <- x[[what]]
-              }
-              x[[what]] <- NULL
-            }
-            x
-          })
-        class(object[["lhoods"]]) <- c("bru_obs_list", "list")
-      }
-
-      object[["inlabru_version"]] <- "2.13.0.9015"
-    }
-
-    object[["inlabru_version"]] <- new_version
-    message(glue("Upgraded bru_info to {new_version}"))
-
-    object_full[["bru_info"]] <- object
-  }
-  object_full
-}
-
-
 #' Methods for bru_info objects
 #'
 #' The `bru_info` class is used to store metadata about `bru` models.
@@ -893,8 +532,10 @@ bru_set_missing.bru_obs <- function(object, keep = FALSE, ...) {
   if (isFALSE(keep)) {
     keep <- -seq_len(bru_response_size(object))
   }
-  bru_set_missing(object[["response_data"]][[object[["response"]]]],
-                  ...) <- keep
+  bru_set_missing(
+    object[["response_data"]][[object[["response"]]]],
+    ...
+  ) <- keep
   object
 }
 
@@ -1614,18 +1255,25 @@ bru_obs_agg <- function(lh,
       .envir = .envir
     )
 
-    if (lh$is_additive) {
-      lh$expr_text <- "BRU_EXPRESSION"
+    pred_text <- lh$pred_expr$pred_text
+    if (is.null(pred_text)) {
+      pred_text <- "BRU_EXPRESSION"
     }
-    lh$expr_text <- glue(
+    pred_text <- glue(
       "{{ibm_eval(BRU_aggregate_mapper, input = BRU_aggregate_input,",
-      " state = {{{lh$expr_text}}})}}"
+      " state = {{{pred_text}}})}}"
     )
-    lh$expr <- parse(text = lh$expr_text)
+    lh$pred_expr <- bru_pred_expr(pred_text,
+      used = lh$pred_expr$used,
+      .envir = lh$pred_expr$.envir
+    )
+    lh$pred_expr$is_additive <- FALSE
+    lh$pred_expr$is_linear <- FALSE
+    lh$pred_expr$is_rowwise <- FALSE
     lh$data_extra[["BRU_aggregate_mapper"]] <- aggregate
     lh$data_extra[["BRU_aggregate_input"]] <- aggregate_input
-    lh$allow_combine <- TRUE
-    lh$is_additive <- FALSE
+
+    lh <- bru_obs_pred_expr_deprecation_fallback(lh)
   }
 
   lh
@@ -1826,7 +1474,6 @@ bru_obs_family_cp_sp <- function(lh, options, .envir) {
   response_data[["BRU_weights"]] <- 1L
 
   if (identical(options[["bru_compress_cp"]], TRUE)) {
-    allow_combine <- TRUE
     new_response_data <- tibble::tibble(
       BRU_E = c(
         rep(0, length(N_data[N_data > 0])),
@@ -1853,13 +1500,14 @@ bru_obs_family_cp_sp <- function(lh, options, .envir) {
         n_block = n_block
       )
 
-    if (lh$is_additive) {
-      lh$expr_text <- "BRU_EXPRESSION"
+    pred_text <- lh$pred_expr$pred_text
+    if (is.null(pred_text)) {
+      pred_text <- "BRU_EXPRESSION"
     }
-    lh$expr_text <- glue("
+    pred_text <- glue("
         {{
           BRU_eta <- {{
-            {lh$expr_text}
+            {pred_text}
           }}
           if (length(BRU_eta) == 1L) {{
             BRU_eta <- rep(BRU_eta, length(BRU_aggregate))
@@ -1871,7 +1519,14 @@ bru_obs_family_cp_sp <- function(lh, options, .envir) {
                 BRU_point_weights[BRU_aggregate])[BRU_cp_block_subset],
             BRU_eta[!BRU_aggregate])
         }}")
-    lh$expr <- parse(text = lh$expr_text)
+    old_pred_expr <- lh$pred_expr
+    lh$pred_expr <- bru_pred_expr(pred_text,
+      used = lh$pred_expr$used,
+      .envir = lh$pred_expr$.envir
+    )
+    lh$pred_expr$is_additive <- FALSE
+    lh$pred_expr$is_linear <- old_pred_expr$is_linear
+    lh$pred_expr$is_rowwise <- FALSE
 
     data <- extended_bind_rows(
       dplyr::bind_cols(data,
@@ -1929,11 +1584,12 @@ bru_obs_family_cp_sp <- function(lh, options, .envir) {
   }
 
   lh$data <- data
-  lh$allow_combine <- TRUE
   lh$response_data <- new_response_data
   lh$response <- "BRU_response"
   lh$inla.family <- "poisson"
   lh$integration_info$ips <- NULL
+
+  lh <- bru_obs_pred_expr_deprecation_fallback(lh)
 
   lh
 }
@@ -1954,12 +1610,11 @@ bru_obs_family_cp <- function(lh, options, .envir) {
   orig_response_data <- lh[["BRU_original_response_data"]]
   response_data <- lh[["response_data"]]
   data <- lh[["data"]]
-  formula <- lh[["formula"]]
   domain <- lh[["integration_info"]][["domain"]]
   samplers <- lh[["integration_info"]][["samplers"]]
   ips <- lh[["integration_info"]][["ips"]]
 
-  response_expr_text <- as.character(formula)[2]
+  response_expr_text <- lh$pred_expr$resp_text
 
   # Catch and handle special cases:
   if (is.null(response) || !inherits(response, "list")) {
@@ -2090,7 +1745,6 @@ bru_obs_family_cp <- function(lh, options, .envir) {
   response_data[["BRU_weights"]] <- 1L
 
   if (identical(options[["bru_compress_cp"]], TRUE)) {
-    allow_combine <- TRUE
     new_response_data <- tibble::tibble(
       BRU_E = c(
         rep(0, length(N_data[N_data > 0])),
@@ -2117,13 +1771,14 @@ bru_obs_family_cp <- function(lh, options, .envir) {
         n_block = n_block
       )
 
-    if (lh$is_additive) {
-      lh$expr_text <- "BRU_EXPRESSION"
+    pred_text <- lh$pred_expr$pred_text
+    if (is.null(pred_text)) {
+      pred_text <- "BRU_EXPRESSION"
     }
-    lh$expr_text <- glue("
+    pred_text <- glue("
         {{
           BRU_eta <- {{
-            {lh$expr_text}
+            {pred_text}
           }}
           if (length(BRU_eta) == 1L) {{
             BRU_eta <- rep(BRU_eta, length(BRU_aggregate))
@@ -2135,7 +1790,14 @@ bru_obs_family_cp <- function(lh, options, .envir) {
                 BRU_point_weights[BRU_aggregate])[BRU_cp_block_subset],
             BRU_eta[!BRU_aggregate])
         }}")
-    lh$expr <- parse(text = lh$expr_text)
+    old_pred_expr <- lh$pred_expr
+    lh$pred_expr <- bru_pred_expr(pred_text,
+      used = lh$pred_expr$used,
+      .envir = lh$pred_expr$.envir
+    )
+    lh$pred_expr$is_additive <- FALSE
+    lh$pred_expr$is_linear <- old_pred_expr$is_linear
+    lh$pred_expr$is_rowwise <- FALSE
 
     data <- extended_bind_rows(
       dplyr::bind_cols(data,
@@ -2184,11 +1846,12 @@ bru_obs_family_cp <- function(lh, options, .envir) {
   }
 
   lh$data <- data
-  lh$allow_combine <- TRUE
   lh$response_data <- new_response_data
   lh$response <- "BRU_response"
   lh$inla.family <- "poisson"
   lh$integration_info$ips <- NULL
+
+  lh <- bru_obs_pred_expr_deprecation_fallback(lh)
 
   lh
 }
@@ -2198,7 +1861,7 @@ check_used_deprecation <- function(used,
                                    include,
                                    exclude,
                                    include_latent,
-                                   formula) {
+                                   expr_text) {
   if (lifecycle::is_present(include) ||
     lifecycle::is_present(exclude) ||
     lifecycle::is_present(include_latent)) {
@@ -2275,7 +1938,8 @@ check_used_deprecation <- function(used,
       )
     }
     if (is.null(used)) {
-      used <- bru_used(formula,
+      used <- bru_used(
+        expr_text,
         effect = include,
         effect_exclude = exclude,
         latent = include_latent
@@ -2285,163 +1949,106 @@ check_used_deprecation <- function(used,
   used
 }
 
-bru_obs_handle_allow_combine <- function(allow_combine,
-                                         data,
-                                         response_data,
-                                         aggregate) {
-  if (!is.logical(allow_combine)) {
+bru_obs_handle_is_rowwise <- function(pred_expr,
+                                      is_rowwise = NULL,
+                                      allow_combine = NULL,
+                                      data,
+                                      response_data,
+                                      aggregate) {
+  if (is.null(is_rowwise) && is.logical(allow_combine)) {
+    is_rowwise <- !allow_combine
+  }
+  if (!is.logical(is_rowwise)) {
     if (!is.null(aggregate)) {
-      allow_combine <- TRUE
+      is_rowwise <- FALSE
     } else if (!is.null(response_data)) {
       bru_log_warn(
         paste0(
           "Non-null response data supplied; ",
-          "guessing allow_combine=TRUE.",
-          "\n  Specify allow_combine explicitly to avoid this warning."
+          "guessing is_rowwise=FALSE.",
+          "\n  Specify is_rowwise explicitly to avoid this warning."
         )
       )
-      allow_combine <- TRUE
+      is_rowwise <- FALSE
     } else if (is.list(data) && !is.data.frame(data)) {
       bru_log_warn(
         paste0(
           "Non data-frame list-like data supplied; ",
-          "guessing allow_combine=TRUE.\n",
-          "  Specify allow_combine explicitly to avoid this warning."
+          "guessing is_rowwise=FALSE.\n",
+          "  Specify is_rowwise explicitly to avoid this warning."
         )
       )
-      allow_combine <- TRUE
+      is_rowwise <- FALSE
     } else {
-      allow_combine <- FALSE
+      is_rowwise <- TRUE
     }
   }
-  allow_combine
+  if (is.null(pred_expr$is_rowwise)) {
+    pred_expr$is_rowwise <- is_rowwise
+  } else {
+    pred_expr$is_rowwise <- pred_expr$is_rowwise && is_rowwise
+  }
+  pred_expr
 }
 
 
 
-
-check_used_deprecation <- function(used,
-                                   include,
-                                   exclude,
-                                   include_latent,
-                                   formula) {
-  if (lifecycle::is_present(include) ||
-    lifecycle::is_present(exclude) ||
-    lifecycle::is_present(include_latent)) {
-    if (!lifecycle::is_present(include)) {
-      include <- NULL
-    } else {
-      bru_log_message(
-        paste0(
-          "The `include` argument of `bru_obs()` is deprecated ",
-          "since inlabru 2.11.0 and may be ignored.\n\t",
-          "If auto-detection doesn't work, use ",
-          "`used = bru_used(effect = include)` instead."
-        ),
-        verbosity = 1L
-      )
-      lifecycle::deprecate_warn(
-        "2.11.0",
-        "bru_obs(include)",
-        "bru_obs(used)",
-        c(
-          "The provided `include` value may be ignored.",
-          "If auto-detection doesn't work, use `bru_used(effect = include)`"
-        )
-      )
+bru_pred_expr <- function(x, used = NULL, .envir = parent.frame()) {
+  resp_text <- NULL
+  if (inherits(x, "formula")) {
+    formula_text <- deparse1(x, collapse = "\n")
+    x <- as.character(x)
+    pred_text <- x[length(x)]
+    if (length(x) > 2) {
+      resp_text <- x[2]
     }
-    if (!lifecycle::is_present(exclude)) {
-      exclude <- NULL
-    } else {
-      bru_log_message(
-        paste0(
-          "The `exclude` argument of `bru_obs()` is deprecated ",
-          "since inlabru 2.11.0 and may be ignored.\n\t",
-          "If auto-detection doesn't work, use ",
-          "`used = bru_used(effect_exclude = exclude)` instead."
-        ),
-        verbosity = 1L
-      )
-      lifecycle::deprecate_warn(
-        "2.11.0",
-        "bru_obs(exclude)",
-        "bru_obs(used)",
-        c(
-          "The provided `exclude` value may be ignored.",
-          paste0(
-            "If auto-detection doesn't work, ",
-            "use `bru_used(effect_exclude = exclude)`"
-          )
-        )
-      )
-    }
-    if (!lifecycle::is_present(include_latent)) {
-      include_latent <- NULL
-    } else {
-      bru_log_message(
-        paste0(
-          "The `include_latent` argument of `bru_obs()` is deprecated ",
-          "since inlabru 2.11.0 and may be ignored.\n\t",
-          "If auto-detection doesn't work, use ",
-          "`used = bru_used(latent = include_latent)` instead."
-        ),
-        verbosity = 1L
-      )
-      lifecycle::deprecate_warn(
-        "2.11.0",
-        "bru_obs(include_latent)",
-        "bru_obs(used)",
-        c(
-          "The provided `include_latent` value may be ignored.",
-          paste0(
-            "If auto-detection doesn't work, ",
-            "use `bru_used(latent = include_latent)`"
-          )
-        )
-      )
-    }
-    if (is.null(used)) {
-      used <- bru_used(formula,
-        effect = include,
-        effect_exclude = exclude,
-        latent = include_latent
-      )
-    }
+  } else {
+    pred_text <- x
+    formula_text <- glue::glue("~ {pred_text}")
   }
-  used
+  is_additive <- bru_is_additive(pred_text)
+  is_additive_dot <- identical(pred_text, ".")
+  if (!is_additive_dot) {
+    pred_expr <- parse(text = pred_text)
+  } else {
+    pred_text <- NULL
+    pred_expr <- NULL
+  }
+  if (is.null(resp_text)) {
+    resp_expr <- NULL
+  } else {
+    resp_expr <- parse(text = resp_text)
+  }
+  structure(
+    list(
+      formula_text = formula_text,
+      pred_text = pred_text,
+      pred_expr = pred_expr,
+      is_additive = is_additive,
+      # Also depends on component defs, so may be changed later:
+      is_linear = is_additive,
+      is_rowwise = NULL,
+      used = if (is.null(used)) bru_used(pred_text) else used,
+      .envir = .envir,
+      resp_text = resp_text,
+      resp_expr = resp_expr
+    ),
+    class = "bru_pred_expr"
+  )
 }
 
-bru_obs_handle_allow_combine <- function(allow_combine,
-                                         data,
-                                         response_data,
-                                         aggregate) {
-  if (!is.logical(allow_combine)) {
-    if (!is.null(aggregate)) {
-      allow_combine <- TRUE
-    } else if (!is.null(response_data)) {
-      bru_log_warn(
-        paste0(
-          "Non-null response data supplied; ",
-          "guessing allow_combine=TRUE.",
-          "\n  Specify allow_combine explicitly to avoid this warning."
-        )
-      )
-      allow_combine <- TRUE
-    } else if (is.list(data) && !is.data.frame(data)) {
-      bru_log_warn(
-        paste0(
-          "Non data-frame list-like data supplied; ",
-          "guessing allow_combine=TRUE.\n",
-          "  Specify allow_combine explicitly to avoid this warning."
-        )
-      )
-      allow_combine <- TRUE
-    } else {
-      allow_combine <- FALSE
-    }
-  }
-  allow_combine
+bru_obs_pred_expr_deprecation_fallback <- function(lh) {
+  # return(lh)
+  lh$is_additive <- lh$pred_expr$is_additive
+  # Also depends on component defs, so may be changed later:
+  lh$linear <- lh$pred_expr$is_linear
+  lh$expr_text <- lh$pred_expr$pred_text
+  lh$expr <- lh$pred_expr$pred_expr
+  lh$used <- lh$pred_expr$used
+  lh$allow_combine <- !lh$pred_expr$is_rowwise
+  lh
 }
+
 
 
 #' @title Observation model construction for usage with [bru()]
@@ -2515,10 +2122,12 @@ bru_obs_handle_allow_combine <- function(allow_combine,
 #' latent vectors are made available to the predictor evaluation is defined by
 #' `bru_used(formula)`, which will include all effects and latent vectors
 #' used by the predictor expression.
-#' @param allow_combine logical; If `TRUE`, the predictor expression may involve
-#'   several rows of the input data to influence the same row. When `NULL`,
-#'   defaults to `FALSE`, unless `response_data` is non-`NULL`, or `data` is a
-#'   `list`, or the likelihood construction requires it.
+#' @param is_rowwise,allow_combine logical; If `is_rowwise` is `FALSE`, the
+#'   predictor expression may involve several rows of the input data to
+#'   influence the same row. When `NULL`, it defaults to `TRUE`, unless
+#'   `response_data` is non-`NULL`, or `data` is a `list`, or the likelihood
+#'   construction requires it. The `allow_combine` argument is a deprecated and
+#'   inverted version, such that `is_rowwise = !allow_combine`.
 #' @param aggregate character ("none" or a valid name for the `type` argument of
 #'   [bm_aggregate()]) or an aggregation `bru_mapper` object ([bm_aggregate()],
 #'   [bm_logsumexp()], or [bm_logitaverage()]). Default `NULL`, interpreted as
@@ -2578,7 +2187,7 @@ bru_obs <- function(formula = . ~ .,
                     samplers = NULL,
                     ips = NULL,
                     used = NULL,
-                    allow_combine = NULL,
+                    is_rowwise = NULL,
                     aggregate = NULL,
                     aggregate_input = NULL,
                     control.family = NULL,
@@ -2588,9 +2197,17 @@ bru_obs <- function(formula = . ~ .,
                     .envir = parent.frame(),
                     include = deprecated(),
                     exclude = deprecated(),
-                    include_latent = deprecated()) {
+                    include_latent = deprecated(),
+                    allow_combine = deprecated()) {
   options <- bru_call_options(options)
   bru_options_set_local(options, .reset = TRUE)
+
+  check_sp_data_deprecation(
+    data = data,
+    response_data = response_data,
+    samplers = samplers,
+    ips = ips
+  )
 
   # Some defaults
   inla.family <- family
@@ -2605,44 +2222,32 @@ bru_obs <- function(formula = . ~ .,
     )
   }
 
+  # Only for deprecated include/exclude/include_latent argument handling:
   formula_char <- as.character(formula)
-
-  # Does the likelihood formula imply an additive predictor?
-  is_additive <- bru_is_additive(formula_char[length(formula_char)])
-  is_additive_dot <- formula_char[length(formula_char)] == "."
-
-  # If not additive, set predictor expression according to the formula's RHS
-  if (!is_additive) {
-    expr_text <- formula_char[length(formula_char)]
-    expr <- parse(text = expr_text)
-  } else {
-    expr_text <- NULL
-    expr <- NULL
-  }
-
-  check_sp_data_deprecation(
-    data = data,
-    response_data = response_data,
-    samplers = samplers,
-    ips = ips
-  )
   used <- check_used_deprecation(
     used = used,
     include = rlang::maybe_missing(include),
     exclude = rlang::maybe_missing(exclude),
     include_latent = rlang::maybe_missing(include_latent),
-    formula = formula
+    expr_text = formula_char[length(formula_char)]
+  )
+
+  # Build expressions from formula:
+  pred_expr <- bru_pred_expr(formula,
+    used = used,
+    .envir = .envir
   )
 
   # Set the response name
-  if (length(formula_char) < 3) {
-    stop("Missing response variable names")
+  if (is.null(pred_expr$resp_text)) {
+    stop("Missing response variable name or expression")
   }
-  response_expr <- rlang::parse_expr(formula_char[2])
+  response_expr <- rlang::parse_expr(pred_expr$resp_text)
+  data_list <- list(response_data = response_data, data = data)
   response <- tryCatch(
     expr = bru_eval_in_data_context(
       !!response_expr,
-      data = list(response_data = response_data, data = data),
+      data = data_list,
       default = NULL,
       .envir = .envir
     ),
@@ -2653,33 +2258,35 @@ bru_obs <- function(formula = . ~ .,
 
   E <- bru_eval_in_data_context(
     {{ E }},
-    data = list(response_data = response_data, data = data),
+    data = data_list,
     default = options[["E"]],
     .envir = .envir
   )
   Ntrials <- bru_eval_in_data_context(
     {{ Ntrials }},
-    data = list(response_data = response_data, data = data),
+    data = data_list,
     default = options[["Ntrials"]],
     .envir = .envir
   )
   weights <- bru_eval_in_data_context(
     {{ weights }},
-    data = list(response_data = response_data, data = data),
+    data = data_list,
     default = 1,
     .envir = .envir
   )
   scale <- bru_eval_in_data_context(
     {{ scale }},
-    data = list(response_data = response_data, data = data),
+    data = data_list,
     default = 1,
     .envir = .envir
   )
 
   data_extra <- as.list(data_extra)
 
-  # Decide on allow_combine default
-  allow_combine <- bru_obs_handle_allow_combine(
+  # Decide on is_rowwise default
+  pred_expr <- bru_obs_handle_is_rowwise(
+    pred_expr = pred_expr,
+    is_rowwise = is_rowwise,
     allow_combine = allow_combine,
     data = data,
     response_data = response_data,
@@ -2701,7 +2308,7 @@ bru_obs <- function(formula = . ~ .,
   lh <- structure(
     list(
       family = family,
-      formula = formula,
+      pred_expr = pred_expr,
       response_data = response_data,
       data = data,
       data_extra = data_extra,
@@ -2710,14 +2317,8 @@ bru_obs <- function(formula = . ~ .,
         domain = domain,
         samplers = samplers
       ),
-      is_additive = is_additive,
-      linear = is_additive, # Not quite correct, as depends on component defs
-      expr_text = expr_text,
-      expr = expr,
       response = "BRU_response",
       inla.family = inla.family,
-      used = used,
-      allow_combine = allow_combine,
       control.family = control.family,
       control.gcpo = control.gcpo,
       tag = tag,
@@ -2725,6 +2326,9 @@ bru_obs <- function(formula = . ~ .,
     ),
     class = "bru_obs"
   )
+
+  lh$formula <- formula
+  lh <- bru_obs_pred_expr_deprecation_fallback(lh)
 
   if (!is.null(aggregate)) {
     lh <- bru_obs_agg(lh,
@@ -3079,10 +2683,11 @@ summary.bru_obs <- function(object, verbose = TRUE, ...) {
       family = object[["family"]],
       data_class = class(object[["data"]]),
       response_class = class(object[["response_data"]][[object[["response"]]]]),
-      predictor = deparse(object[["formula"]]),
-      is_additive = object[["is_additive"]],
-      is_linear = object[["linear"]],
-      used = object[["used"]],
+      predictor = glue::glue("{object$pred_expr$formula_text}"),
+      is_additive = object[["pred_expr"]][["is_additive"]],
+      is_linear = object[["pred_expr"]][["is_linear"]],
+      is_rowwise = object[["pred_expr"]][["is_rowwise"]],
+      used = bru_used(object),
       tag = object[["tag"]]
     ),
     class = "summary_bru_obs"
@@ -3153,7 +2758,7 @@ print.summary_bru_obs <- function(x, ...) {
       "    Data class: {data_class}\n",
       "    Response class: {response_class}\n",
       "    Predictor: {predictor}\n",
-      "    Additive/Linear: {is_additive}/{is_linear}\n",
+      "    Additive/Linear/Rowwise: {is_additive}/{is_linear}/{is_rowwise}\n",
       "    Used components: {format(lh$used)}",
       .trim = FALSE,
       tag = if (is.null(lh$tag) || is.na(lh$tag)) {