+++ tests/testthat/test-aggregate.R
@@ -36,7 +36,7 @@ test_that("Aggregated Gaussian observations, using fm_block_eval", {
           )
         )
       ),
-      allow_combine = TRUE
+      is_rowwise = FALSE
     )
   )
 
@@ -69,7 +69,7 @@ test_that("Aggregated Gaussian observations, using fm_block_eval", {
       control.family = list(
         hyper = list(prec = list(initial = 6, fixed = TRUE))
       ),
-      allow_combine = TRUE
+      is_rowwise = FALSE
     )
   )
 
@@ -346,7 +346,7 @@ test_that("Aggregated Poisson observations, using mapper", {
       family = "poisson",
       response_data = obs,
       data = pred,
-      allow_combine = TRUE
+      is_rowwise = FALSE
     )
   )
 
@@ -375,7 +375,7 @@ test_that("Aggregated Poisson observations, using mapper", {
       E = E,
       response_data = obs,
       data = pred,
-      allow_combine = TRUE
+      is_rowwise = FALSE
     )
   )
 
@@ -425,7 +425,7 @@ test_that("Aggregated Gaussian observations, using mapper", {
           )
         )
       ),
-      allow_combine = TRUE
+      is_rowwise = FALSE
     )
   )
 
@@ -456,7 +456,7 @@ test_that("Aggregated Gaussian observations, using mapper", {
       control.family = list(
         hyper = list(prec = list(initial = 6, fixed = TRUE))
       ),
-      allow_combine = TRUE
+      is_rowwise = FALSE
     )
   )