+++ tests/testthat/test-data.R
@@ -9,7 +9,7 @@ test_that("Component construction: default mesh/mapping, data is list", {
     formula = y ~ .,
     data = list(x = c(1, 1.5, 2, NA, 3, 4), y = 11:15),
     used = bru_used(effect = "effect"),
-    allow_combine = TRUE
+    is_rowwise = FALSE
   )
 
   cmp1 <- bru_comp_list(~ effect(c(1, 1.5, 2, 3, 4), model = "iid") - 1)
@@ -69,7 +69,7 @@ test_that("Component construction: unsafe intercepts, data is list", {
   cmp <- bru_comp_list(~ something_unknown - 1)
   lik <- bru_obs(
     formula = response ~ ., data = list(response = 1:5),
-    allow_combine = TRUE
+    is_rowwise = FALSE
   )
   lik <- bru_used_update(lik, labels = names(cmp))
   expect_error(
@@ -93,14 +93,14 @@ test_that("Component construction: separate response_data input", {
     formula = y ~ c(sum(effect), sum(effect^2)),
     data = list(x = c(1, 1.5, 2, 3, 4)),
     response_data = data.frame(y = c(11, 12)),
-    allow_combine = TRUE
+    is_rowwise = FALSE
   )
 
   lik2 <- bru_obs("gaussian",
     formula = y ~ c(sum(effect), sum(effect^2)),
     data = data.frame(x = c(1, 1.5, 2, 3, 4)),
     response_data = data.frame(y = c(11, 12)),
-    allow_combine = TRUE
+    is_rowwise = FALSE
   )